+++ Plan/bukkit/src/main/java/com/djrapitops/plan/system/listeners/bukkit/DeathEventListener.java
