--- Plan/common/src/main/java/com/djrapitops/plan/modules/server/bukkit/BukkitServerPropertiesModule.java
@@ -1,40 +1,0 @@
-/*
- *  This file is part of Player Analytics (Plan).
- *
- *  Plan is free software: you can redistribute it and/or modify
- *  it under the terms of the LGNU Lesser General Public License v3 as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  Plan is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  LGNU Lesser General Public License for more details.
- *
- *  You should have received a copy of the GNU Lesser General Public License
- *  along with Plan. If not, see <https://www.gnu.org/licenses/>.
- */
-package com.djrapitops.plan.modules.server.bukkit;
-
-import com.djrapitops.plan.Plan;
-import com.djrapitops.plan.system.info.server.properties.BukkitServerProperties;
-import com.djrapitops.plan.system.info.server.properties.ServerProperties;
-import dagger.Module;
-import dagger.Provides;
-
-import javax.inject.Singleton;
-
-/**
- * Dagger module for Bukkit ServerProperties.
- *
- * @author Rsl1122
- */
-@Module
-public class BukkitServerPropertiesModule {
-
-    @Provides
-    @Singleton
-    ServerProperties provideServerProperties(Plan plugin) {
-        return new BukkitServerProperties(plugin.getServer());
-    }
-}
