+++ Plan/common/src/main/java/com/djrapitops/plan/modules/server/sponge/SpongeSuperClassBindingModule.java
@@ -21,7 +21,7 @@ import com.djrapitops.plan.system.database.SpongeDBSystem;
 import com.djrapitops.plan.system.importing.EmptyImportSystem;
 import com.djrapitops.plan.system.importing.ImportSystem;
 import com.djrapitops.plan.system.info.server.ServerInfo;
-import com.djrapitops.plan.system.info.server.SpongeServerInfo;
+import com.djrapitops.plan.system.info.server.ServerServerInfo;
 import com.djrapitops.plan.system.listeners.ListenerSystem;
 import com.djrapitops.plan.system.listeners.SpongeListenerSystem;
 import com.djrapitops.plan.system.settings.config.ConfigSystem;
@@ -43,8 +43,8 @@ public class SpongeSuperClassBindingModule {
 
     @Provides
     @Singleton
-    ServerInfo provideSpongeServerInfo(SpongeServerInfo spongeServerInfo) {
-        return spongeServerInfo;
+    ServerInfo provideSpongeServerInfo(ServerServerInfo serverServerInfo) {
+        return serverServerInfo;
     }
 
     @Provides
