--- Plan/common/src/main/java/com/djrapitops/plan/system/info/server/SpongeServerInfo.java
@@ -1,39 +1,0 @@
-/*
- *  This file is part of Player Analytics (Plan).
- *
- *  Plan is free software: you can redistribute it and/or modify
- *  it under the terms of the LGNU Lesser General Public License v3 as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  Plan is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  LGNU Lesser General Public License for more details.
- *
- *  You should have received a copy of the GNU Lesser General Public License
- *  along with Plan. If not, see <https://www.gnu.org/licenses/>.
- */
-package com.djrapitops.plan.system.info.server;
-
-import com.djrapitops.plan.system.database.DBSystem;
-import com.djrapitops.plan.system.info.server.properties.ServerProperties;
-import com.djrapitops.plan.system.settings.config.PlanConfig;
-import com.djrapitops.plan.system.webserver.WebServer;
-import dagger.Lazy;
-
-import javax.inject.Inject;
-
-public class SpongeServerInfo extends BukkitServerInfo {
-
-    @Inject
-    public SpongeServerInfo(
-            ServerProperties serverProperties,
-            ServerInfoFile serverInfoFile,
-            DBSystem dbSystem,
-            Lazy<WebServer> webServer,
-            PlanConfig config
-    ) {
-        super(serverProperties, serverInfoFile, dbSystem, webServer, config);
-    }
-}
